+++ web/src/AppShell.tsx
@@ -2,12 +2,12 @@ import { App } from "./App";
 import { CoreNavigationProvider } from "./CoreNavigation";
 import { FederatedRuntimeProvider } from "./federatedRuntime";
 import { HostRegistryProvider } from "./hostRegistry";
-import { coreSurfaceRegistry } from "./surfaceRegistry";
+import { productSurfaceRegistry } from "./productAssembly";
 
 export function AppShell() {
   return (
     <HostRegistryProvider>
-      <CoreNavigationProvider registry={coreSurfaceRegistry}>
+      <CoreNavigationProvider registry={productSurfaceRegistry}>
         <FederatedRuntimeProvider>
           <App />
         </FederatedRuntimeProvider>